💄 stretch card images when necessary
Because the minimum card width is set to 350px and the container padding
is 30px we need a width of 760px to display two cards.

Reversely at 759px a container can only contain one card. The width of
this card will be 699px. This means content editors should upload images
of minimum 699px to prevent this stretching.